pyproject.toml
[project]
name = "fasthtml-test-tailwind"
version = "0.1.0"
description = "fasthtml test with tailwind"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.24",
]
